admin-case-feature

+++ Domain/CRCIS.Web.INoor.CRM.Domain/Cases/ImportCase/Commands/MoveCaseToCurrentAdminCardboardCommand.cs
@@ -8,10 +8,10 @@ namespace CRCIS.Web.INoor.CRM.Domain.Cases.ImportCase.Commands
 {
    public class MoveCaseToCurrentAdminCardboardCommand
     {
-        public long AdminId { get;private set; }
+        public int AdminId { get;private set; }
         public long CaseId { get;private set; }
 
-        public MoveCaseToCurrentAdminCardboardCommand(long adminId, long caseId)
+        public MoveCaseToCurrentAdminCardboardCommand(int adminId, long caseId)
         {
             AdminId = adminId;
             CaseId = caseId;
