+++ Domain/CRCIS.Web.INoor.CRM.Domain/Cases/PendingCase/Commands/MoveCaseToPartnerAdminCardboardCommand.cs
@@ -8,10 +8,10 @@ namespace CRCIS.Web.INoor.CRM.Domain.Cases.PendingCase.Commands
 {
   public  class MoveCaseToPartnerAdminCardboardCommand
     {
-        public long AdminId { get; private set; }
+        public int AdminId { get; private set; }
         public long CaseId { get; private set; }
 
-        public MoveCaseToPartnerAdminCardboardCommand(long adminId, long caseId)
+        public MoveCaseToPartnerAdminCardboardCommand(int adminId, long caseId)
         {
             AdminId = adminId;
             CaseId = caseId;
